+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.2.13 (21-02-2026)
+- Added guild tabard customization to character tool.
 - Added housing decor categories to quickly filter decor models.
 - Fixed an issue that prevented certain animations from exporting correctly.
 - Fixed an issue that would prevent certain models from loading/exporting.